moved text changed listeners
